ENTITY 로직에 대한 테스트 코드를 제외하고 격리 // TempTestCode
